+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Now prefabs can be linked to any entity within scene's hierarchy.
+
 ### Changed
 
 ### Fixed